pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mp-psrp-rl"
version = "0.1.0"
description = "Modular reinforcement learning stack for the multi-product petrol station replenishment problem."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "gymnasium>=0.29",
    "torch>=2.3",
    "hydra-core>=1.3",
    "tensorboard>=2.15",
    "matplotlib>=3.8",
    "pandas>=2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
